Extract MsaglSvgRenderer: deduplicate shared SVG rendering from Arrow and Vertex graph serializers
