+++ src/Zametek.ViewModel.ProjectPlan/GraphManagement/GraphSerializer/IMsaglSvgRenderer.cs
@@ -1,0 +1,9 @@
+using Zametek.Common.ProjectPlan;
+
+namespace Zametek.ViewModel.ProjectPlan
+{
+    public interface IMsaglSvgRenderer
+    {
+        byte[] RenderToSvg(Microsoft.Msagl.Drawing.Graph graph, BaseTheme theme);
+    }
+}
